refactor: bring the branch back inside the architecture contract
`scripts/check_architecture.py` now reports 0 errors, as `main` does. The
remaining 21 violations are closed by moving responsibility wherever a
cohesive unit existed, and by a reviewed baseline raise only where the
addition is irreducible.

Moved out (each with a docstring saying why it is its own owner):

  - `workflows/bundle_import_evidence.py` -- the three questions the
    unresolved-import detectors share, held apart: did a version-compatible
    sibling provide this symbol in OLD (the only one that licenses the
    BREAKING kind), did OLD carry this import at all, and do the outward
    DT_NEEDED edges all resolve to system providers. Conflating them at the
    call site is what produced both of this area's bugs, in opposite
    directions.
  - `workflows/bundle_unresolved_audit.py` -- the single-bundle audit
    detector, which deliberately keeps a different emptiness policy: it has
    no previous release to point at, so no outward edges is no evidence.
  - `policy/persistent_hygiene.py` -- the verdict exclusion and its two
    limits.
  - `report/kind_rollup.py` -- both halves of the Markdown per-kind rollup,
    with the `compute_`/`render_` split preserved inside it.
  - `extract/header_exclusions.py` -- `--exclude-header`'s matching rule.
  - `compat/run_inputs.py`, `compat/multi_library_run.py`,
    `compat/descriptor_expansion.py`.
  - `tests/test_bundle_unresolved_imports.py` (533 lines),
    `tests/test_persistent_hygiene_verdict.py` (276 lines).

Baselines raised, each with its own rationale in `architecture/debt.yaml`:
the `--exclude-header` operand threaded through three layers, one
`ChangeKindMeta` catalog entry, one `_TYPE_BEARING_DISCRIMINATOR_KINDS`
membership, one `BugClass` registry entry, the rollup's three call sites and
one dataclass field, and the shared `_removal_pair` bundle fixture. A
catalog entry, a registry entry and a dataclass field have no owner other
than their catalog, registry and dataclass.

The `cli-contract` allowlist is re-pinned to `compat/cli.py`'s new line
numbers. Its three Tier-1 call sites stay in that file on purpose: the gate
scans `cli*.py`/`compat/cli.py`, so moving one would take it out of the
gate's scope rather than out of the line budget.

Tests that patched a moved helper are repointed at its new owner, per
AGENTS.md's "import from the owner, and patch the owner".

This commit is unsigned for the same reason as the last: the environment's
SSH signer was removed mid-session and the configured key file is empty.
