+++ abicheck/checker.py
@@ -131,8 +131,6 @@ from .policy.classification import (
     RISK_KINDS as _RISK_KINDS,
     Verdict,
     compute_verdict,
-    excluded_from_verdict_as_persistent_hygiene,
-    policy_kind_sets,
 )
 from .policy.disposition_close import finalize_ledger
 from .policy.disposition_ledger import (
@@ -140,6 +138,7 @@ from .policy.disposition_ledger import (
     record_suppressed_change,
 )
 from .policy.evidence_status import EvidenceTier, is_cross_source_resolved
+from .policy.persistent_hygiene import drop_persistent_hygiene
 from .policy.policy_file_namespaces import (
     experimental_namespaces as _experimental_namespaces,
     internal_namespaces as _internal_namespaces,
@@ -207,84 +206,10 @@ def _compute_verdict_for(
     if stage is not None:
         stage.classify(all_unsuppressed)
         all_unsuppressed = evaluated_for_policy(all_unsuppressed)
-    all_unsuppressed = _drop_persistent_hygiene(all_unsuppressed, policy, policy_file)
+    all_unsuppressed = drop_persistent_hygiene(all_unsuppressed, policy, policy_file)
     if policy_file is not None:
         return policy_file.compute_verdict(all_unsuppressed)
     return compute_verdict(all_unsuppressed, policy=policy)
-
-
-def _drop_persistent_hygiene(
-    changes: list[Change],
-    policy: str,
-    policy_file: PolicyFile | None,
-) -> list[Change]:
-    """*changes* minus the pre-existing cross-source hygiene debt that must
-    not drive this release's verdict
-    (:func:`~abicheck.policy.classification.
-    excluded_from_verdict_as_persistent_hygiene`).
-
-    Applied here, at the one verdict chokepoint, rather than in
-    :func:`_verdict_scored_population` alongside the ``RESOLVED`` exclusion:
-    deciding this needs the finding's *resolved* category, which is a
-    function of the active policy, and only this function has it. The
-    category is resolved against the same kind sets the verdict computation
-    itself is about to use -- ``policy_file.base_policy`` when a policy
-    document is in effect, so a document that changed the base profile does
-    not get judged against ``strict_abi``'s partitions.
-
-    **A finding the policy document states a rule about is never excluded**,
-    whether through ``overrides:`` or ``reclassify:``. Those rules are applied
-    by ``PolicyFile.compute_verdict``, not by the base kind sets, so the
-    category this function resolves does not see them: a project that
-    deliberately set ``exported_not_public: breaking`` would have had every
-    persistent instance silently dropped from its own verdict -- the exact
-    opposite of what it asked for. Verified by executing that case, not by
-    reasoning about it; an earlier revision of this docstring asserted it
-    "fails in the safe direction" and was simply wrong
-    (``TestPersistentHygieneRespectsAnExplicitPolicy``).
-
-    The exclusion is deliberately withheld for *any* mention of the kind,
-    including one that lowers its severity, rather than only for a
-    promotion. A project that has written the kind into its policy at all is
-    managing that kind itself, and quietly removing its findings from the
-    verdict is not this function's call to make.
-    """
-    base = policy_file.base_policy if policy_file is not None else policy
-    sets = policy_kind_sets(base)
-    return [
-        c
-        for c in changes
-        if _policy_governs(c, policy_file)
-        or not excluded_from_verdict_as_persistent_hygiene(c, *sets)
-    ]
-
-
-def _policy_governs(change: Change, policy_file: PolicyFile | None) -> bool:
-    """Whether *policy_file* states a rule that applies to *change*.
-
-    Both rule namespaces, because both change a finding's resolved verdict
-    and neither is visible in the base profile's kind sets:
-    ``overrides:`` is kind-global, so a kind lookup answers it; ``reclassify:``
-    is selector-scoped, so the rule is *asked whether it matches this
-    finding* rather than mined for the kinds it mentions.
-
-    Asking the rule is not a refinement, it is the only correct form. An
-    earlier version collected ``rule.kinds``/``rule.kind`` -- neither of
-    which exists; the field is ``change_kind`` -- so every reclassify rule
-    was silently ignored, and a rule like ``{kind: exported_not_public,
-    symbol: x, to: break}`` had its finding dropped before
-    ``PolicyFile.compute_verdict`` ever saw it, turning a deliberate
-    ``break`` into ``NO_CHANGE`` (Codex review). Mining kind names would
-    also have missed a rule with no kind selector at all -- a symbol- or
-    namespace-scoped rule that legitimately applies to a hygiene finding.
-    """
-    if policy_file is None:
-        return False
-    if change.kind in (getattr(policy_file, "overrides", {}) or {}):
-        return True
-    return any(
-        rule.matches(change) for rule in (getattr(policy_file, "reclassify", ()) or ())
-    )
 
 
 def _verdict_scored_population(
